+++ src/main/java/com/clickhouse/kafka/connect/RawClickHouseSinkConnector.java
@@ -7,7 +7,6 @@ import org.apache.kafka.common.config.Config;
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.connect.connector.Task;
 import org.apache.kafka.connect.sink.SinkConnector;
-
 import org.apache.kafka.connect.sink.SinkConnectorContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;