+++ src/main/java/com/clickhouse/kafka/connect/sink/ClickHouseSinkConfig.java
@@ -50,6 +50,10 @@ public class ClickHouseSinkConfig {
     public static final String SOURCE = "source";
     public static final String DB_TYPE = "dbType";
     public static final String ID_FIELD = "idField";
+    public static final String TOLERATE_STATE_MISMATCH = "tolerateStateMismatch";
+    public static final String BYPASS_SCHEMA_VALIDATION = "bypassSchemaValidation";
+    public static final String BYPASS_FIELD_CLEANUP = "bypassFieldCleanup";
+    public static final String IGNORE_PARTITIONS_WHEN_BATCHING = "ignorePartitionsWhenBatching";
 
     public static final int MILLI_IN_A_SEC = 1000;
     private static final String databaseDefault = "default";
@@ -63,12 +67,6 @@ public class ClickHouseSinkConfig {
     public static final Integer tableRefreshIntervalDefault = 0;
     public static final Boolean exactlyOnceDefault = Boolean.FALSE;
     public static final Boolean customInsertFormatDefault = Boolean.FALSE;
-    public enum StateStores {
-        NONE,
-        IN_MEMORY,
-        REDIS,
-        KEEPER_MAP
-    }
 
     private final String hostname;
     private final int port;
@@ -98,6 +96,10 @@ public class ClickHouseSinkConfig {
     private final String source;
     private final String dbType;
     private final String idField;
+    private final boolean tolerateStateMismatch;
+    private final boolean bypassSchemaValidation;
+    private final boolean bypassFieldCleanup;
+    private final boolean ignorePartitionsWhenBatching;
 
     public enum InsertFormats {
         NONE,
@@ -273,6 +275,10 @@ public class ClickHouseSinkConfig {
             }
         }
         this.clientVersion = props.getOrDefault(CLIENT_VERSION, "V1");
+        this.tolerateStateMismatch = Boolean.parseBoolean(props.getOrDefault(TOLERATE_STATE_MISMATCH, "false"));
+        this.bypassSchemaValidation = Boolean.parseBoolean(props.getOrDefault(BYPASS_SCHEMA_VALIDATION, "false"));
+        this.bypassFieldCleanup = Boolean.parseBoolean(props.getOrDefault(BYPASS_FIELD_CLEANUP, "false"));
+        this.ignorePartitionsWhenBatching = Boolean.parseBoolean(props.getOrDefault(IGNORE_PARTITIONS_WHEN_BATCHING, "false"));
 
         LOGGER.debug("ClickHouseSinkConfig: hostname: {}, port: {}, database: {}, username: {}, sslEnabled: {}, timeout: {}, retry: {}, exactlyOnce: {}",
                 hostname, port, database, username, sslEnabled, timeout, retry, exactlyOnce);
@@ -295,6 +301,7 @@ public class ClickHouseSinkConfig {
     private static ConfigDef createConfigDef() {
         ConfigDef configDef = new ConfigDef();
 
+        //TODO: At some point we should group these more clearly
         String group = "Connection";
         int orderInGroup = 0;
         configDef.define(HOSTNAME,
@@ -567,6 +574,46 @@ public class ClickHouseSinkConfig {
                 ++orderInGroup,
                 ConfigDef.Width.SHORT,
                 "Client version"
+        );
+        configDef.define(TOLERATE_STATE_MISMATCH,
+                ConfigDef.Type.BOOLEAN,
+                false,
+                ConfigDef.Importance.LOW,
+                "Tolerate state mismatch. default: false",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "Tolerate state mismatch."
+        );
+        configDef.define(BYPASS_SCHEMA_VALIDATION,
+                ConfigDef.Type.BOOLEAN,
+                false,
+                ConfigDef.Importance.LOW,
+                "Bypass schema validation. default: false",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "Bypass schema validation."
+        );
+        configDef.define(BYPASS_FIELD_CLEANUP,
+                ConfigDef.Type.BOOLEAN,
+                false,
+                ConfigDef.Importance.LOW,
+                "Bypass field cleanup. default: false",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "Bypass field cleanup."
+        );
+        configDef.define(IGNORE_PARTITIONS_WHEN_BATCHING,
+                ConfigDef.Type.BOOLEAN,
+                false,
+                ConfigDef.Importance.LOW,
+                "Ignore partitions when batching. Flag ignored when exactlyOnce=true. default: false",
+                group,
+                ++orderInGroup,
+                ConfigDef.Width.SHORT,
+                "Ignore partitions when batching."
         );
         return configDef;
     }